requirements.txt
numpy
pillow
matplotlib
tqdm
opencv-python